+++ arnav-portfolio/src/components/ui/skiper-ui/skiper19.jsx
@@ -15,11 +15,11 @@ const Skiper19 = () => {
       className="mx-auto flex h-[350vh] w-screen flex-col items-center overflow-hidden bg-[#FAFDEE] px-4 text-[#1F3A4B]"
     >
       <div className="mt-42 relative flex w-fit flex-col items-center justify-center gap-5 text-center">
-        <h1 className="font-jakarta-sans relative z-10 text-7xl font-medium tracking-[-0.08em] lg:text-9xl">
+        <h1 className="font-jakarta-sans relative z-10 text-7xl font-normal tracking-[-0.08em] lg:text-9xl">
           The Stroke <br /> That follows the <br />
           Scroll Progress
         </h1>
-        <p className="font-jakarta-sans relative z-10 max-w-2xl text-xl font-medium text-[#1F3A4B]">
+        <p className="font-jakarta-sans relative z-10 max-w-2xl text-xl font-normal text-[#1F3A4B]">
           Scroll down to see the effect
         </p>
 
@@ -30,10 +30,10 @@ const Skiper19 = () => {
       </div>
 
       <div className="rounded-4xl font-jakarta-sans w-full translate-y-[200vh] bg-[#1F3A4B] pb-10 text-[#FAFDEE]">
-        <h1 className="mt-10 text-center text-[15.5vw] font-bold leading-[0.9] tracking-tighter lg:text-[16.6vw]">
+        <h1 className="mt-10 text-center text-[15.5vw] font-normal leading-[0.9] tracking-tighter lg:text-[16.6vw]">
           skiperui.com
         </h1>
-        <div className="mt-80 flex w-full flex-col items-start gap-5 px-4 font-medium lg:mt-0 lg:flex-row lg:justify-between">
+        <div className="mt-80 flex w-full flex-col items-start gap-5 px-4 font-normal lg:mt-0 lg:flex-row lg:justify-between">
           <div className="flex w-full items-center justify-between gap-12 uppercase lg:w-fit lg:justify-center">
             <p className="w-fit text-sm">
               punjab, india <br />